requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
bertopic==0.15.0
umap-learn==0.5.4
hdbscan==0.8.33
sentence-transformers==2.2.2
jieba==0.42.1
openpyxl==3.1.2
python-docx==0.8.11
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.0
requests==2.31.0
gunicorn==21.2.0
